# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Materials Management API with Golang",
        "title": "Materials Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Bao Tran",
            "url": "http://github.com/remiehneppo",
            "email": "bao.tran080898@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "/api/v1",
    "paths": {
...
        "/materials-request": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new material request with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material-requests"
                ],
                "summary": "Create a new material request",
                "parameters": [
                    {
                        "description": "Material request data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateMaterialRequestReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Material request created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/materials-request/cancel/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel an existing material request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material-requests"
                ],
                "summary": "Cancel a material request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Material Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Material request canceled successfully",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request - ID is required",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/materials-request/export": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Export a material request to a downloadable DOCX document",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
                ],
                "tags": [
                    "material-requests"
                ],
                "summary": "Export material request to DOCX",
                "parameters": [
                    {
                        "description": "Export request data",
                        "name": "export",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.MaterialRequestExport"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DOCX file download",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/materials-request/filter": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve material requests based on filter criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material-requests"
                ],
                "summary": "Filter material requests",
                "parameters": [
                    {
                        "description": "Filter criteria for material requests",
                        "name": "filter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.MaterialRequestFilter"
                        }
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Material requests filtered successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.MaterialRequest"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/materials-request/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the details of an existing material request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material-requests"
                ],
                "summary": "Update a material request",
                "parameters": [
                    {
                        "description": "Material request update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.MaterialRequestUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Material request updated successfully",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/materials-request/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific material request using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material-requests"
                ],
                "summary": "Get material request by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Material Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Material request retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.MaterialRequest"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request - ID is required",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/materials-request/{id}/issue": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atomically assigns the next request number and adds requested quantities to Material Profile Reality",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material-requests"
                ],
                "summary": "Issue a draft material request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Material Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Material request issued",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.IssueMaterialRequestResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Session is not authorized",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "403": {
                        "description": "Only the requester can issue",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "404": {
                        "description": "Material request not found",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "409": {
                        "description": "Material request is not a draft",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.CreateMaterialRequestReq": {
            "type": "object",
            "required": [
                "materials_for_equipment",
                "sector"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "maintenance_instance_id": {
                    "type": "string"
                },
                "materials_for_equipment": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/types.MaterialsForEquipment"
                    }
                },
                "sector": {
                    "type": "string"
                }
            }
        },
...
        "types.IssueMaterialRequestResponse": {
            "type": "object",
            "properties": {
                "issued_at": {
                    "type": "integer"
                },
                "issued_by": {
                    "type": "string"
                },
                "request_number": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "types.Material": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "types.MaterialRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "issued_at": {
                    "type": "integer"
                },
                "issued_by": {
                    "type": "string"
                },
                "maintenance_instance_id": {
                    "type": "string"
                },
                "materials_for_equipment": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/types.MaterialsForEquipment"
                    }
                },
                "num_of_request": {
                    "type": "integer"
                },
                "requested_at": {
                    "type": "integer"
                },
                "requested_by": {
                    "type": "string"
                },
                "requester_user_id": {
                    "type": "string"
                },
                "sector": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "types.MaterialRequestExport": {
            "type": "object",
            "required": [
                "material_request_id"
            ],
            "properties": {
                "material_request_id": {
                    "type": "string"
                }
            }
        },
        "types.MaterialRequestFilter": {
            "type": "object",
            "properties": {
                "equipment_machinery_id": {
                    "type": "string"
                },
                "maintenance_instance_id": {
                    "type": "string"
                },
                "num_of_request": {
                    "type": "integer"
                },
                "requested_at_end": {
                    "type": "integer"
                },
                "requested_at_start": {
                    "type": "integer"
                },
                "requested_by": {
                    "type": "string"
                },
                "sector": {
                    "type": "string"
                }
            }
        },
        "types.MaterialRequestUpdate": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "materials_for_equipment": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/types.MaterialsForEquipment"
                    }
                },
                "sector": {
                    "type": "string"
                }
            }
        },
        "types.MaterialsForEquipment": {
            "type": "object",
            "properties": {
                "consumable_supplies": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/types.Material"
                    }
                },
                "replacement_materials": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/types.Material"
                    }
                }
            }
        },
...
        "types.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}